refactor: move every stylesheet as sibling of their counterpart
