+++ src/frontend/index.scss
@@ -1,11 +1,11 @@
-@import './fonts.scss';
-@import './HtmlEditor.scss';
-@import './Editor.scss';
-@import './AuthDialog.scss';
-@import './FeedbackButton.scss';
-@import './DiscordButton.scss';
-@import './MainMenuLabel.scss';
-@import './CustomEmbeddableRenderer.scss';
+@font-face {
+    font-family: 'Roboto';
+    src: url('/assets/fonts/Roboto-VariableFont_wdth,wght.ttf') format('truetype-variations');
+    font-weight: 100 900;
+    font-stretch: 75% 100%;
+    font-style: normal;
+    font-display: swap;
+}
 
 /* Override Excalidraw styles */
