+++ src/frontend/src/CustomEmbeddableRenderer.scss
