+++ src/frontend/src/pad/controls/StateIndicator.scss
