+++ src/frontend/src/pad/editors/Editor.scss
