+++ src/frontend/src/pad/editors/HtmlEditor.scss
