--- src/frontend/src/pad/styles/index.scss
@@ -1,6 +1,0 @@
-@import './ActionButton.scss';
-@import './ActionButtonGrid.scss';
-@import './ControlButton.scss';
-@import './StateIndicator.scss';
-@import './Dashboard.scss';
-@import './CanvasBackups.scss';
