--- src/frontend/src/styles/fonts.scss
@@ -1,9 +1,0 @@
-/* Font declarations */
-@font-face {
-  font-family: 'Roboto';
-  src: url('/assets/fonts/Roboto-VariableFont_wdth,wght.ttf') format('truetype-variations');
-  font-weight: 100 900;
-  font-stretch: 75% 100%;
-  font-style: normal;
-  font-display: swap;
-}
