+++ src/frontend/src/ui/AuthDialog.scss
