+++ src/frontend/src/ui/BackupsDialog.scss
