+++ src/frontend/src/ui/MainMenu.scss
@@ -1,0 +1,24 @@
+.main-menu-label {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+    margin-top: 0.25rem;
+    margin-bottom: 0.25rem;
+    font-weight: 500;
+    font-size: 1rem;
+    cursor: default;
+  }
+  
+  .main-menu-top-row {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding-left: 0.5rem;
+  }
+  
+  .main-menu-label span:last-child {
+    color: #fa8933;
+    font-weight: 1000;
+  }
+
